scripts/test_setup/requirements.txt
# Зависимости для скриптов настройки тестового окружения

# PostgreSQL adapter
psycopg2-binary>=2.9.0

# YAML parser для конфигов
PyYAML>=6.0

# Генератор тестовых данных (русская локаль)
Faker>=19.0.0
